+++ docs/articles/analysis.R
@@ -20,14 +20,14 @@ wastdr::wastdr_setup()
 if (file.exists("~/tracks.Rda")){
     load("~/tracks.Rda")
 } else {
-    track_records <- wastdr::get_wastd("turtle-nest-encounters")
+    track_records <- wastdr::wastd_GET("turtle-nest-encounters")
     save_file <- Sys.getenv("WASTDR_SAVE_LOCALLY", unset = FALSE)
     if (save_file==TRUE){
         save(track_records, file = "~/track_records.Rda")
         load("~/track_records.Rda")
     }
     # listviewer::jsonedit(utils::head(track_records$features))
-    tracks <- parse_turtle_nest_encounters(track_records)
+    tracks <- wastdr::parse_turtle_nest_encounters(track_records)
     if (save_file == TRUE){
         save(tracks, file = "~/tracks.Rda")
     }
@@ -77,11 +77,16 @@ add_lookups <- . %>%
     left_join(species_colours, by="species") %>%
     left_join(nest_type_text, by="nest_type")
 
-filter_2017 <- . %>% dplyr::filter(date > dmy("17/11/2017")) %>% add_lookups
+filter_2017 <- . %>% dplyr::filter(date > dmy("01/10/2017"), date < dmy("31/03/2018"))
 filter_broome <- . %>% dplyr::filter(area_name=="Cable Beach Broome")
+filter_cbb1 <- . %>% dplyr::filter(site_name=="Cable Beach Broome Sector 1")
+filter_cbb2 <- . %>% dplyr::filter(site_name=="Cable Beach Broome Sector 2")
+filter_cbb3 <- . %>% dplyr::filter(site_name=="Cable Beach Broome Sector 3")
+
 filter_eighty_mile_beach <- . %>% dplyr::filter(area_name=="Eighty Mile Beach Caravan Park")
 filter_anna_plains <- . %>% dplyr::filter(area_name=="Anna Plains")
-filter_port_hedland <- . %>% dplyr::filter(site_name=="Port Hedland Turtle Nesting Beaches")
+filter_port_hedland_cemetery <- . %>% dplyr::filter(site_name=="Port Hedland Cemetery Beach")
+filter_port_hedland_prettypool <- . %>% dplyr::filter(site_name=="Port Hedland Pretty Pool Beach")
 filter_west_pilbara <- . %>% dplyr::filter(area_name=="West Pilbara Turtle Program beaches Wickam")
 filter_delambre <- . %>% dplyr::filter(area_name=="Delambre Island")
 filter_rosemary <- . %>% dplyr::filter(area_name=="Rosemary Island")
@@ -89,7 +94,7 @@ filter_thevenard <- . %>% dplyr::filter(area_name=="Thevenard Island")
 
 ## ---- eval=T-------------------------------------------------------------
 tracks_map <- function(track_data) {
-    l <- leaflet(width=800, height=600) %>% 
+    l <- leaflet(width=1000, height=800) %>% 
         addProviderTiles("Esri.WorldImagery", group = "Aerial") %>%
         addProviderTiles("OpenStreetMap.Mapnik", group = "Place names") %>%
         clearBounds()
@@ -120,6 +125,13 @@ tracks_map <- function(track_data) {
 }
 
 ## ----data_wp_daily-------------------------------------------------------
+species_by_type <- . %>% 
+  filter(nest_age=="fresh") %>%
+  group_by(species, nest_type) %>% 
+  tally() %>%
+  ungroup() %>% 
+  tidyr::spread(nest_type, n, fill=0)
+
 daily_species_by_type <- . %>% 
     filter(nest_age=="fresh") %>%
     group_by(date, species, nest_type) %>% 
@@ -139,6 +151,7 @@ tracks_ts <- . %>%
             # ggplot2::geom_line() +
             ggplot2::scale_x_date(breaks = scales::pretty_breaks(),
                                   labels = scales::date_format("%d %b %Y")) +
+            ggplot2::scale_y_continuous(limits = c(0, NA)) +
             ggplot2::xlab("Date") +
             ggplot2::ylab("Number counted per day") +
             ggplot2::ggtitle("Nesting activity") +
@@ -172,12 +185,22 @@ tracks_emb %>% DT::datatable(.)
 tracks_emb %>% daily_summary
 tracks_emb %>% tracks_ts
 
-## ---- fig.width=7, fig.height=5------------------------------------------
-tracks_pth <- tracks %>% filter_2017 %>% filter_port_hedland
-tracks_pth %>% tracks_map
-tracks_pth %>% DT::datatable(.)
-tracks_pth %>% daily_summary
-tracks_pth %>% tracks_ts
+## ---- fig.width=9, fig.height=5------------------------------------------
+tracks_pth_cem <- tracks %>% filter_2017 %>% filter_port_hedland_cemetery
+tracks_pth_ppo <- tracks %>% filter_2017 %>% filter_port_hedland_prettypool
+
+
+tracks_pth_cem %>% tracks_map
+tracks_pth_cem %>% DT::datatable(.)
+tracks_pth_cem %>% daily_summary
+tracks_pth_cem %>% tracks_ts %T>% ggsave(filename="~/pth_daily_tracks_cem.png", device="png", width=9, height=5)
+
+
+tracks_pth_ppo %>% tracks_map
+tracks_pth_ppo %>% DT::datatable(.)
+tracks_pth_ppo %>% daily_summary
+tracks_pth_ppo %>% tracks_ts %T>% ggsave(filename="~/pth_daily_tracks_ppo.png", device="png", width=9, height=5)
+
 
 # named_nests_pth <- tracks_pth %>% filter(!(is.na(name)))
 # named_nests_pth %>% tracks_map