+++ docs/articles/getting-wastd.R
@@ -8,7 +8,7 @@ library(DT)
 library(ggplot2)
 
 ## ----call_api, eval=FALSE------------------------------------------------
-#  turtle_nest_encounters <- get_wastd('turtle-nest-encounters')
+#  turtle_nest_encounters <- wastd_GET('turtle-nest-encounters')
 #  listviewer::jsonedit(turtle_nest_encounters$content, width = 800, height = 600)
 #  tracks <- parse_turtle_nest_encounters(turtle_nest_encounters)
 
@@ -32,4 +32,13 @@ tracks <- parse_turtle_nest_encounters(turtle_nest_encounters)
 DT::datatable(animals)
 # DT::datatable(nests)
 # DT::datatable(tracks)
+
+## ----surveys, fig.height=5, fig.width=10---------------------------------
+survey_res <- wastd_GET("surveys")
+surveys <- survey_res %>% parse_surveys()
+
+surveys_pth <- surveys %>% dplyr::filter(site_id %in% c(35, 45)) 
+
+# survey_ts_plot <- . %>% dplyr::group_by(date, site_name) %>% dplyr::tally %>%
+# ggplot(., aes(date, n, fill=site_name)) + ggplot2::geom_tile() + theme_minimal() + labs(title="Surveys", fill="Site", y="Number of surveys", x="Date")