+++ tests/testthat/test_data.R
@@ -17,15 +17,15 @@ testthat::test_that("wastd_api_response prints", {
   data(animal_encounters)
   capture.output(print(animal_encounters))
 
-  data("turtle_nest_encounters_hatched")
-  capture.output(print(turtle_nest_encounters_hatched))
+  data("nests")
+  capture.output(print(nests))
 })
 
 
 testthat::test_that("turtle_nest_encounters_hatched parses correctly to nests", {
-  data(turtle_nest_encounters_hatched)
+  data(tne)
   data(nests)
-  fresh_nests <- parse_turtle_nest_encounters(turtle_nest_encounters_hatched)
+  fresh_nests <- parse_turtle_nest_encounters(tne)
   testthat::expect_equal(nrow(fresh_nests), nrow(nests))
   # Compare pickled and fresh nests excluding list columns (like obs)
   testthat::expect_equal(
@@ -35,13 +35,13 @@ testthat::test_that("turtle_nest_encounters_hatched parses correctly to nests", 
 })
 
 testthat::test_that("turtle_nest_encounters parses correctly to tracks", {
-  data(turtle_nest_encounters)
-  data(tracks)
-  fresh_tracks <- parse_turtle_nest_encounters(turtle_nest_encounters)
-  testthat::expect_equal(nrow(fresh_tracks), nrow(tracks))
+  data(tne)
+  data(nests)
+  fresh_tracks <- parse_turtle_nest_encounters(tne)
+  testthat::expect_equal(nrow(fresh_tracks), nrow(nests))
   # Compare pickled and fresh nests excluding list columns (like obs)
   testthat::expect_equal(
     fresh_tracks %>% dplyr::select(-obs, -photos),
-    tracks %>% dplyr::select(-obs, -photos)
+    nests %>% dplyr::select(-obs, -photos)
   )
 })