+++ NEWS.md
@@ -4,5 +4,11 @@
 * Package size: 9.4Mb - will split out example data once 
   package matures and release frequency slows down. 
 
+## 1.0.4
+
+* WIP: parse Encounter.survey; wastd_POST.
+* Changed example dataset names to "tne" and "nests"
+* WIP: add helper methods to filter and summarise data
+
 ## 1.0.3
 * AnimalEncounters: add `maturity`